# Avoid unnecessarily refreshing `SliderBodyPiece`'s path