This reduces the draw thread overhead when many sliders are selected in the editor. I'm making this a very local optimisation for now because I want to avoid changing all the `UpdateFrom` flow (profiling that alone shows that it has quite low overhead relative to other issues).

Before (on https://github.com/ppy/osu/pull/20965): Update 44fps, Draw 22fps
After: Update 44fps, Draw 38fps